+++ packages/lavas-core-vue/core/config-reader.js
@@ -114,7 +114,10 @@ export default class ConfigReader {
     constructor(cwd, env, customConfigPath) {
         this.cwd = cwd;
         this.env = env;
-        this.customConfigPath = customConfigPath;
+        // fix https://github.com/lavas-project/lavas/issues/50
+        if (customConfigPath !== 'dev') {
+            this.customConfigPath = customConfigPath;
+        }
     }
 
     /**