+++ packages/lavas-core-vue/test/fixtures/vuetify/core/service-worker.js
@@ -1,0 +1,25 @@
+/**
+ * @file service-worker.js with workbox api
+ * @desc [example](https://workbox-samples.glitch.me/examples/workbox-sw/)
+ * @author *__ author __*{% if: *__ email __* %}(*__ email __*){% /if %}
+ */
+
+/**
+ * Import workbox-sw with `importScripts` function.
+ * 1. Need adding publicPath manually.
+ * 2. If the version of workbox updated, modification is also required.
+ */
+importScripts('/static/js/workbox-sw.prod.v2.1.2.js');
+
+const workboxSW = new WorkboxSW({
+    cacheId: 'lavas-cache',
+    ignoreUrlParametersMatching: [/^utm_/],
+    skipWaiting: true,
+    clientsClaim: true
+});
+
+// Define precache injection point.
+workboxSW.precache([]);
+
+// Define response for HTML request.
+workboxSW.router.registerNavigationRoute('/appshell/main');