+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 =======
 
-## Lavas-core-vue [1.0.3-rc.1] - 2018-1-17
+## Lavas-core-vue [1.0.4-rc.1] - 2018-1-17
 
 ### Changed
 
 - [Feature] Allow developers use custom variables in `core/index.html.tmpl`
+- [Fix] Developers with old version of lavas (≤ 2.2.2) can use `lavas dev` correctly.
 
 ## Lavas-core-vue [1.0.2-rc.0] - 2018-1-10